pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "heinrich-triz"
version = "0.1.0"
description = "AI-powered TRIZ (Theory of Inventive Problem Solving) system for systematic innovation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Nick Scherbakov", email = "contact@heinrich-triz.org"}
]
maintainers = [
    {name = "Nick Scherbakov", email = "contact@heinrich-triz.org"}
]
keywords = [
    "TRIZ",
    "innovation",
    "problem-solving",
    "artificial-intelligence",
    "engineering",
    "systematic-innovation",
    "inventive-principles",
    "contradiction-matrix"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "PyYAML>=6.0",
    "numpy>=1.21.0",
    "dataclasses-json>=0.5.0",
    "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pylint>=2.15.0",
    "isort>=5.11.0",
    "pre-commit>=3.0.0"
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0"
]
ui = [
    "rich>=12.0.0",
    "colorama>=0.4.0"
]

[project.urls]
Homepage = "https://github.com/NickScherbakov/Heinrich-The-Inventing-Machine"
Documentation = "https://github.com/NickScherbakov/Heinrich-The-Inventing-Machine/docs"
Repository = "https://github.com/NickScherbakov/Heinrich-The-Inventing-Machine"
Issues = "https://github.com/NickScherbakov/Heinrich-The-Inventing-Machine/issues"

[project.scripts]
heinrich = "heinrich_cli:main"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# mypy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "llm: marks tests that require LLM access"
]

# Coverage configuration
[tool.coverage.run]
source = ["heinrich"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]

# Pylint configuration
[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
]

[tool.pylint.format]
max-line-length = 100

# Flake8 configuration (in setup.cfg as flake8 doesn't support pyproject.toml)
